add strength changes

+++ database/migrations/2021_02_10_220239_create_teams_table.php
@@ -16,7 +16,7 @@ class CreateTeamsTable extends Migration
         Schema::create('teams', function (Blueprint $table) {
             $table->id();
             $table->text('team_name');
-            $table->integer('strength')->default(0);
+            $table->float('strength')->default(0);
         });
     }
